+++ src/Nodes/Channel/MidSide.h
@@ -14,7 +14,7 @@ public:
     std::shared_ptr<Output> getSide() const;
 
 protected:
-    void setNumSamplesNoLock(int numSamples) override;
+    void setNumSamplesNoLock(size_t numSamples) override;
     void processNoLock() override;
 
 private:
